Bug solved to some extent
Bug stated in previous commit is solved to some extent by
filtering the line to the condition that y coordinate of
both the ends of the line should be greater than half of
image height that is the line should completly lie in lower
half of the image. Another approach can also be used in which
the image will be cropped for the top leaving the lower half
for lines to be detected as the road is always present below
the horizon which is generally near the half of the image.